requirements.txt
# # Web framework
# Flask==2.3.3
# Flask-Cors==4.0.0
# gunicorn==21.2.0

# # Database
# SQLAlchemy==2.0.20
# Flask-SQLAlchemy==3.1.1
# alembic==1.12.0

# # Image processing and OCR
# opencv-python==4.8.0.76
# pytesseract==0.3.10
# Pillow==10.0.0
# pdf2image==1.16.3
# numpy==1.25.2

# # NLP and text processing
# nltk==3.8.1
# scikit-learn==1.3.0
# spacy==3.6.1
# transformers==4.32.1

# # Authentication
# Flask-JWT-Extended==4.5.2
# bcrypt==4.0.1

# # Utilities
# python-dotenv==1.0.0
# Werkzeug==2.3.7


# Web framework
Flask
Flask-Cors
gunicorn

# Database
SQLAlchemy
Flask-SQLAlchemy
alembic

# Image processing and OCR
opencv-python
pytesseract
Pillow
pdf2image
numpy

# NLP and text processing
nltk
scikit-learn
spacy
transformers

# Authentication
Flask-JWT-Extended
bcrypt

# Utilities
python-dotenv
Werkzeug
